requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
langchain-core
langchainhub
selenium
webdriver-manager
lxml